+++ src/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <link rel="stylesheet" href="./style.css">
+    <link rel="stylesheet" href="path/to/font-awesome/css/font-awesome.min.css">
     <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
     <title>Document</title>
   </head>
@@ -34,12 +35,5 @@
     <script type="module" src="router.js"></script>
 
   </body>
-  <footer>
-    <ul class="footer">
-      <a href="https://github.com/BiancaSherika" target="_blank"><strong>@ Bianca Sherika</strong></a>
-      <a href="https://github.com/ly-cardozo" target="_blank"><strong>@ Eliane Cardozo</strong></a>
-      <a href="https://github.com/gabrielasilva1991" target="_blank"><strong>@ Gabriela Silva</strong></a>
-      <a href="https://www.laboratoria.la/" target="_blank"><strong>by Laboratória</strong></a>
-    </ul>
-  </footer>
+  
 </html>
