+++ src/pages/profile/index.js
@@ -1,0 +1,9 @@
+
+export const Profile = () => {
+    const profileCreated = document.createElement('div');
+    profileCreated.innerHTML = `
+     <h1>Profile teste</h1>
+    
+    `
+    return profileCreated
+  };
